Move settings into namespace
